TempFolder.DeleteAfter

+++ CSharpExt/Utility/TempFolder.cs
@@ -10,25 +10,31 @@ namespace Noggog.Utility
     public class TempFolder : IDisposable
     {
         public DirectoryInfo Dir { get; private set; }
+        public bool DeleteAfter = true;
 
-        public TempFolder()
+        public TempFolder(bool deleteAfter = true)
         {
             this.Dir = new DirectoryInfo(Path.Combine(Path.GetTempPath(), Path.GetRandomFileName()));
             this.Dir.Create();
+            this.DeleteAfter = deleteAfter;
         }
 
-        public TempFolder(DirectoryInfo dir)
+        public TempFolder(DirectoryInfo dir, bool deleteAfter = true)
         {
             this.Dir = dir;
             if (!dir.Exists)
             {
                 this.Dir.Create();
             }
+            this.DeleteAfter = deleteAfter;
         }
 
         public void Dispose()
         {
-            this.Dir.DeleteEntireFolder();
+            if (DeleteAfter)
+            {
+                this.Dir.DeleteEntireFolder();
+            }
         }
     }
 }
